updadate user route class

+++ lib/a-token.class.php
@@ -10,15 +10,14 @@ class AToken
 
     public function register(User $user)
     {
-        $user->update(['atoken' => (int)(config()->get(TOKEN_REGISTER))]);
+        $user->update(['atoken' => config()->get(TOKEN_REGISTER)]);
         $this->log(
             userIdx : $user-> profile()[IDX],
             reason: 'tokenRegister',
             pointAfterApply:  $user -> profile()[POINT],
-            tokenApply: (int)(config()->get(TOKEN_REGISTER)),
-            tokenAfterApply : (int)(config()->get(TOKEN_REGISTER)),
+            tokenApply: config()->get(TOKEN_REGISTER),
+            tokenAfterApply : config()->get(TOKEN_REGISTER),
         );
-
     }
 
     public function accountDelete():array{
